requirements.txt
Django===5.0.6
djangorestframework===3.15.1

# Authentication
djoser===2.2.2
djangorestframework_simplejwt===5.3.1

# CORS
django-cors-headers===4.3.1

# Upload Images
Pillow===10.3.0

# Environment
django-environ===0.11.2

# Password Encoding
argon2-cffi===23.1.0

# PostgreSQL driver
psycopg2-binary===2.9.9

# Swagger UI
drf-spectacular===0.27.2

# Testing
factory-boy===3.3.0
coverage===7.5.3
